navi_line: better logging and added blur support
Added a blur orthogonal to the filter direction (i.e. horizontal for the
vertical filter and vertical for the horizontal filter). This is equivalent to
using a matched pulse-width filter with height greater than one and is better
than using a standard box filter on the input or output. Eventually, this
should be converted to use real-world coordinates instead of pixel coordinates.

Removed output normalization for the line_filter_hor and line_filter_ver debug
topics. This gives a better picture of the algorithm's performance when there
is no line visible in the image and lets the output be used for tuning the
threshold.